requirements.txt
pyuseragents~=1.0.5
pydantic~=2.4.2
loguru~=0.7.2
PyYAML~=6.0.1
tls_client
